[design] moved all Beagle Core classes in a common package: Beagle core.
