+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/analysis/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Types for Analysers extending Beagle.
  */
-package de.uka.ipd.sdq.beagle.analysis;
+package de.uka.ipd.sdq.beagle.core.analysis;
