+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/judge/package-info.java
@@ -2,4 +2,4 @@
  * Interfaces and classes for final judges, analysers that ultimately decide Beagle’s
  * results.
  */
-package de.uka.ipd.sdq.beagle.judge;
+package de.uka.ipd.sdq.beagle.core.judge;
