+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/measurement/ResourceDemandMeasurementResult.java
@@ -1,4 +1,4 @@
-package de.uka.ipd.sdq.beagle.measurement;
+package de.uka.ipd.sdq.beagle.core.measurement;
 
 import de.uka.ipd.sdq.beagle.core.ResourceDemandingInternalAction;
