+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/measurement/package-info.java
@@ -2,4 +2,4 @@
  * Types for measurement tools, performing run time or counting measurements on running
  * software.
  */
-package de.uka.ipd.sdq.beagle.measurement;
+package de.uka.ipd.sdq.beagle.core.measurement;
